ENH: load standard data set for perf test
The moving volume is in the build directory so everyone can replicate
the issues.  Start slicer and type "Control-5" to trigger the timing
test.  Note that when no volume is displayed, the performance is stable
over time when Control-5 is hit repeatedly.  Now display the 'moving'
volume in the Red slice display and then note that the frames per second
goes down rapidly on repeated trials.